requirements.txt
asgiref==3.3.4
astroid==2.3.3
boto3==1.12.31
botocore==1.15.31
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
defusedxml==0.7.0rc1
dj-database-url==0.5.0
Django==3.2.4
django-appconf==1.0.4
django-compressor==2.4
django-crispy-forms==1.9.1
django-environ==0.4.5
django-heroku==0.3.1
django-sass-processor==0.8
django-simple-history==3.0.0
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
idna==2.9
isort==4.3.21
jmespath==0.9.5
jsonfield==3.1.0
lazy-object-proxy==1.4.3
libsass==0.20.0
mccabe==0.6.1
mrjob==0.7.1
oauthlib==3.1.0
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
pylint==2.4.4
pylint-django==2.0.15
pylint-plugin-utils==0.6
python-dateutil==2.8.1
python-dotenv==0.17.1
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.3.1
rcssmin==1.0.6
requests==2.23.0
requests-oauthlib==1.3.0
rjsmin==1.1.0
rope==0.17.0
s3transfer==0.3.3
six==1.14.0
social-auth-app-django==3.4.0
social-auth-core==3.3.3
sqlparse==0.3.1
urllib3==1.25.8
whitenoise==5.1.0
wrapt==1.11.2
